cratewide: group all public C API in one module
`cassandra.h` contains a lot of C API functions, which comprise
the public API of the CPP Driver. As CPP-Rust driver strives to be
a drop-in replacement for the CPP Driver, it is crucial to be aware
of all the public API functions it exposes, and all `cassandra.h`
functions that are **not** implemented (yet, or are never planned).

This commit creates a module `api`, which re-exports all public C API
of the CPP-Rust driver, grouped by the `cassandra.h` order.
This makes it easier to see which functions are implemented,
as well as to see which functions are missing.

This has another advantage: it allows to use the C API functions in Rust
code, because they are now accessible for imports; before, they were
present in non-`pub` modules, so e.g. integration tests could not use
them.