cluster: set serial consistency to None by default
The default serial consistency in the Rust driver is LocalSerial, which
is different to Any in the CPP driver. This commit sets the default
serial consistency to None, which is equivalent to Any.
This makes CPP-Rust driver consistent with the CPP driver in terms
of the default serial consistency.

Note: Rust driver purposefully set the default serial consistency to
LocalSerial instead of None. The rationale taken from the issue
(https://github.com/scylladb/scylla-rust-driver/issues/277):

> Using lightweight transactions in CQL requires setting an additional
> consistency level, the so called serial consistency level, on top of
> the regular one. The serial consistency level can only take two
> values: SERIAL and LOCAL_SERIAL. There's currently no default, so
> using lwt results in an error message: `Consistency level for LWT is
> missing for a request with conditions`. We should consider picking
> a sensible default value in order to improve the user experience
> of the driver.

Considering this, we should consider changing the default serial
consistency in CPP-Rust driver to LocalSerial, too.